requirements.txt
Flask-Mail
psycopg2-binary
